backend/requirements.txt
alembic==1.13.1
aniso8601==10.0.0
asgiref==3.7.2
async-timeout==5.0.1
blinker==1.9.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-allauth==0.63.3
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
googletrans==4.0.0rc1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.3
lazy_loader==0.4
Mako==1.3.9
MarkupSafe==2.1.4
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.1
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.31.0
rfc3986==1.5.0
scikit-image==0.23.1
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
tifffile==2025.2.18
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
